Move csv->binary conversion reference into library
...under the perspective/csv subpackage, to keep it out of the way of
the core visualization code. Users may want to use an alternative input
source with a different source, requiring a different conversion
process, but keeping this inside the library does provide a ready-made
textual import format which is easy to use with databases and
spreadsheets - and just as easily used as a reference for writing an
importer for some other data format.