Create zero-copy accessor for image pixels
We can get a rendering performance improvement of approximately 10% by
using a zero-copy mechanism for editing image pixel values rather than
using the Go image library's At() and Set() methods, due to a reduction
in memory copying and type conversions.

Done here with the creation of the new accessor and its application to
the scatter visualization (the case for which the performance difference
was tested) to illustrate its use. The remaining visualizations can be
updated to use this in a subsequent commit.