Create reference HTTP server implementation
...minimal HTTP server implementation, to demonstrate use of this
visualization library to generate images for consumption by a browser
interface or other service which can use the HTTP API.

At this phase, the input data is hard-coded, most of the internal error
handling will cause the server to exit without restarting, and no
attempt is made to cache the input data for re-use between requests.

In other words, there is still work to be done here. But at least we can
now load a generated graph by dropping a URL in a browser's address bar.