Separate loading event log from processing events
Break the process of loading a binary log of events out from the
process of recording and rendering them as a visualization - which
were previously done together in feeds.GeneratePNGFromBinLog().

This is done for the sake of (1) allowing the HTTP interface to
effectively cache loaded content by retaining a pointer to the
memory-mapped binary log file, and (2) allowing for future introduction
of alternative data sources (like having the HTTP interface accept
event data and record it in some form of internal data storage structure
or external database).