Cache mapped event-data pointers in HTTP server
Once mapped, our data-source files can remain that way for the life of
the server (or until explicitly removed, once we bring in
data-source-management as a feature of the HTTP interface). This reduces
per-request load on the serving system by a bit (and will become far
more relevant when we start doing things like allowing importation of
ephemerally-stored event data - in which case sharing these mapped
structures and the pointers to them would be a very good idea).

TODO note included for actions needed for the sake of safety when the
HTTP service is modified to handle requests concurrently rather than
serially.