Pre-calculate color increment in sweep vis
This is invariant over event-data-point recordings, so we can compute
it once at the time of the sweep visualization generator's construction
and use the computed value for all subsequent event recordings.

It does seem, from benchmarking comparisons, that the Go compiler is at
least smart enough to make this optimization on its own - but the code
change also clarified for a human reader that this value can be expected
to remain stable through a sweep-visualization generator's lifetime.