Accept time-unit suffixes on offsets into the past
For the perspective server, accept suffixes in the set {year, month,
day, h, m, s} and interpret appropriately for time offsets into the past
relative to "NOW()".

The year unit is assumed to mean 365 days and the month unit is assumed
to mean 31 days. An interface wishing to sanely and precisely handle
special cases can always specify the offsets in days instead.

For any unit other than "s", a post-suffix pluralizer "s" character will
be ignored. Thus "year" and "years" are treated as semantically
equivalent suffixes.

Seconds are assumed if no unit is given.

As stated in the comment inline with this code submission, while it is
possible for an off-label use to run these visualizations with a
different concept of minimum time granularity and when "zero time" is,
such usage should avoid the human-frieldly and suffixed time
specifications as the results may not match with expectations.