Account for unfinished events in ribbon
Show unfinished events as grayscale vertical bands along with the blue
spikes down from the top for successes and the red spikes up from the
bottom for failures.

Also initialize counts to 1 instead of 0 to avoid division-by-zero
errors if we're missing examples of one or more of the event status
cases in our input feed (an issue noticed during testing of this
change).