Implement filtered event-data dump for server
Binary dump of event data, filtered according to the specified time
bounds and event type (if supplied), for consumption by clients which
want this raw data for tabular display or client-size graphical
rendering.

Event data is output as a 20-byte struct with each of the fields in {id,
start_time, run_time, type, status} represented by a 32-bit
little-endian signed integer.

This format is chosen (over the more space-efficient internal
representation) for the sake of simplicity when parsing with things like
Javascript which have a concept of typed binary arrays, but no solid
support for parsing blocks of data as arrays of c-style structs.