Add static-content serving capacility
Useful for hosting such things as dasboards from the same server as is
used to generate the graphs, while leaving the nature of this static
content up to the server's end users.

Static content is served from a /static/ subpath under the server's
root path in the request URL and /var/opt/perspective/static/ as
stored on the server.