Implement event filtering by done/failed/running
Status filtering mask support introduced so the HTTP server and
command-line interface can allow clients to specify if they are
interested in seeing events which are done, failed, in-progress, or some
combination of the above. Default is to show all events which match on
the other filter parameters (event type and start-time window).