Server action to get success rate over time range
Allow clients to use a "success-rate" action to query the a perspective
server for the percentage of successfully-completed events (relative to
all completed events) initiated within any specified time and type
filters given for an event feed managed by the Perspective server.

This is returned to the client as a simple text value (like "99.957%).