Keep in-progress events green in scatter
Don't allow in-progress events to blow out and desaturate to while like
completed events in the scatter visualization.

After watching a sample implementation for a while, it became clear that
it is more helpful to have these stay clearly distinct from completed
events than it is to have the extra density-representation range that
can come from desaturation.