Create run-time-line visualization
Visualization showing trend of average event run-times over time (with
data placement determined by event start time).

Color of line at at point along its length is correlated to the
relative preponderance of successful, failed, and active events. If
filtered to only show one of these kinds of event, the line will be
uniformly blue, red, or green in tone as corresponds to the exit status
which has been filtered for.

Y-axis position of the line is on a logarithmic scale rather than a
linear one, same as is used for the scatter visualization.

No smoothing is applied to this visualization, unlike the count-lines
visualization, because the amount of "jitter" in event run times is
itself important data.