Add lookback option for limiting event-feed scan
Add a "lookback" option to allow clients to specify how far back in an
event feed the scan for plotting event data should begin from. If a
lookback value is not specified (or a value of 0 is given), then the
entire feed will be scanned - as was the case prior to this commit.

Positive and negative lookback values have the same meaning, with the
lookback being the absolute value of the lookback count.

The lookback scan distance may be increased to align the start of the
memory mapping for the binary-log scan to a page boundary. It will also,
of course, be limited by the length of the feed - with any lookback
value larger than the feed's file size being interpreted as instruction
to scan the entire file.

This mechanism is being introduced as a heuristic for speeding up "tail"
scans of a feed which grows over time without requiring that the data be
indexed or fully sorted (though some approximation of chronological
ordering is assumed to be the case for appending data to the log in
order for the use of this feature to make sense).