Fix "bleedaround" issue in rendering scatter plots
Points at the left edge of the scatter visualizations were able to bleed
around to the right edge due to a defect in the edge-detection check
meant to prevent this from happening. Which, while this artefact was
relatively harmless, it could be easily seen in upscaled versions of
the plots and distracted from the intended rendering of the data.

This fixes that.