Interpret lookback arg as event count, not bytes
End consumers should not have to know the size of an event data record
(or the nature of the backing feed at all) in order to use the lookback
heuristic to speed up the rendering of "tail" graphs from a large feed.

This commit changes the interpretation of this value from being a count
of bytes to being a count of events by multiplying the lookback value by
the size of the event data record to get a byte offset from the end of
the file from which to begin reading data.