Increase grid line contrast
Make the grid lines a bit brighter, on feedback that they are not easy
to see under all lighting conditions.

This would be a good thing to consider making a tunable parameter for in
the future, so the grid contrast can be increased or reduced (even to
the point of eliminating the grid) according to user preference.